Refactor code to reduce duplication
