+++ app/Main.hs
@@ -3,7 +3,6 @@
 module Main where
 
 import Args (Args (..), getArgs)
-import Azure (fetchSchedules, getToken)
 import Control.Monad (when)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Text as T
@@ -16,11 +15,12 @@ import Data.Time.LocalTime
     localTimeToUTC,
     timeZoneOffsetString,
   )
-import Entities (Days (..), Room (..), allRooms)
-import Meetings (chooseBestMeeting, getMeetings)
-import Print (infoPrint, prettyPrint)
+import Meet.Azure (fetchSchedules, getToken)
+import Meet.Entities (Days (..), Room (..), allRooms)
+import Meet.Meetings (chooseBestMeeting, getMeetings)
+import Meet.Print (infoPrint, prettyPrint)
+import Meet.Utils
 import System.Exit (exitSuccess)
-import Utils
 
 main :: IO ()
 main = do
