+++ meet.cabal
@@ -17,11 +17,12 @@ extra-doc-files: CHANGELOG.md
 
 library
     exposed-modules:
-        Azure
-        Entities
-        Meetings
-        Print
-        Utils
+        Meet.Args
+        Meet.Azure
+        Meet.Entities
+        Meet.Meetings
+        Meet.Print
+        Meet.Utils
 
     hs-source-dirs:     src
     default-language:   Haskell2010
@@ -45,6 +46,7 @@ library
         unix >= 2.8.4 && < 2.9,
         bytestring >= 0.12.1.0 && < 0.13,
         binary >= 0.8.9.1 && < 0.9,
+        optparse-applicative >= 0.18.1 && < 0.19,
 
 executable meet
     main-is:            Main.hs
